requirements.txt
h5py
psutil
tqdm
termcolor
tensorboard
tensorboardX
imageio
imageio-ffmpeg
egl_probe>=1.0.1
torch
torchvision
gym
matplotlib
seaborn
wandb
scikit-learn
numpy>=1.13.3
numba>=0.52.0
scipy>=1.2.3
free-mujoco-py==2.1.4
Pillow
opencv-python
pynput
